requirements.txt
openai
groq
google-generativeai
cohere
requests
python-dotenv